aact-832: With additional column displayed in Data Dictionary, the table has become a bit more awkward to use - since users need to scroll to the right to see important info.
We have row count appear in that table, which isn’t the most appropriate place to display that bit of info since it refers to # of rows per table, but this displays info about columns.  We have the row count also appear in the ‘table’ table which seems more appropriate.  So removing the row count from Data Dictionary.